executor: remove old, never used "IndexLookUpJoin"
